refactor: split atomicEffects.ts into focused modules
Split the monolithic atomicEffects.ts (440 lines) into domain-focused modules:
- atomicCombatEffects.ts: GainAttack, GainBlock
- atomicResourceEffects.ts: GainMove, GainInfluence, GainMana, GainCrystal
- atomicProgressionEffects.ts: GainFame, ChangeReputation
- atomicCardEffects.ts: DrawCards, GainHealing, TakeWound
- atomicModifierEffects.ts: ApplyModifier
- atomicHelpers.ts: Shared utilities (updatePlayer, updateElementalValue)

Also simplified applyGainAttack by extracting helper functions and
eliminated duplication by sharing elementToPropertyKey with reverse.ts.
