+++ packages/core/src/engine/effects/reverse.ts
@@ -50,27 +50,8 @@ import {
   COMBAT_TYPE_RANGED,
   COMBAT_TYPE_SIEGE,
 } from "../../types/effectTypes.js";
-import { getLevelsCrossed, ELEMENT_FIRE, ELEMENT_ICE, ELEMENT_COLD_FIRE } from "@mage-knight/shared";
-import type { Element } from "@mage-knight/shared";
-import { MIN_REPUTATION, MAX_REPUTATION } from "./atomicEffects.js";
-
-/**
- * Maps Element type to the property key in ElementalAttackValues.
- * Element uses "cold_fire" but ElementalAttackValues uses "coldFire".
- */
-function elementToPropertyKey(element: Element | undefined): "physical" | "fire" | "ice" | "coldFire" {
-  if (!element) return "physical";
-  switch (element) {
-    case ELEMENT_FIRE:
-      return "fire";
-    case ELEMENT_ICE:
-      return "ice";
-    case ELEMENT_COLD_FIRE:
-      return "coldFire";
-    default:
-      return "physical";
-  }
-}
+import { getLevelsCrossed } from "@mage-knight/shared";
+import { MIN_REPUTATION, MAX_REPUTATION, elementToPropertyKey } from "./atomicEffects.js";
 
 // ============================================================================
 // REVERSE EFFECT
